Alterando a Arquitetura, para algo mais escalável.
