+++ .idea/models/app.py
@@ -6,7 +6,7 @@ load_dotenv("D:\Projects\evolution\.env")
 EVOLUTION_AUTH = os.getenv("AUTHENTICATION_API_KEY")
 
 BASE_URL = "http://localhost:8080"
-INSTANCE_NAME="" # o nome da instancia que foi colocada no /manager no site
+INSTANCE_NAME="teste" # o nome da instancia que foi colocada no /manager no site
 
 NUMERO = "5562994980317"
 MENSAGEM = "Testando api"
@@ -15,11 +15,13 @@ headers = {
     'apikey': EVOLUTION_AUTH,
     'Content-Type': 'application/json'
 }
+
 payload = {
     'number': f'{NUMERO}',
     'text': f'{MENSAGEM}',
     # 'delay': 10000, # simular "digitando"
 }
+
 response = requests.post(
     url=f'{BASE_URL}/message/sendText/{INSTANCE_NAME}',
     json=payload,
